Fix DCM Bug and move printing order around.
The Compasss object receives a dcm matrix for orientation and uses it
to calculate the heading angle of the vehicle. This was previously
equivalent to an euler angle reading of (0,0,0) which is correct for
small angles, but not for large angles. The fix is tested on the
airframe and working.